HOST_TOOLS_PREFIX is now deprecated, replace with value.
This change is a workaround to deal with the latest step towards complete
removal of gcc from the build. HOST_TOOLS_PREFIX used to point to a host gcc
toolchain, but now it's gone. The toolchain is still present, however, so this
change works. We'll need a better solution soon as gcc is going to be yanked at
some point. I've opened b/117951707 to track eventual solution.

Change-Id: If5052481a0566e0140ef77b1697d96f7478fefb0